+++ docs/website/publish.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 GIT_USER=alphaville \
-  CURRENT_BRANCH=master \
+  CURRENT_BRANCH=main \
   USE_SSH=true \
   yarn deploy